backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
boto3==1.34.0
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20
langchain-community==0.0.10
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
httpx==0.25.2

